+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 - Removed unused component `EditQuestionPage` to avoid confusion with other components. This is a legacy component. [#379]
 
 ### Fixed
+- Fixed ordering issue for Sections on the Edit Templates page [#436]
 - Fixed missing `as deps` in `Dockerfile.prod`.
 - Fixed wrong node image in `Dockerfile.prd`. Also fixed a number of linting issues.
 - Fixed bug in `/projects/[projectId]/dmp/[dmpId]/download` because of unused `FileIcon` [#376]